refactor(grants): simplify grants page to render RoundCard list only
